pyproject.toml
[project]
name = "api-core"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "adrf>=0.1",
    "app-store-server-library>=1.9",
    "boto3>=1.42",
    "channels>=4.3",
    "channels-redis>=4.3",
    "dj-database-url>=3.1",
    "django>=5.2",
    "django-allauth>=65.13",
    "django-cors-headers>=4.9",
    "django-oauth-toolkit>=3.2",
    "django-redis>=5.4",
    "django-storages[s3]>=1.14",
    "djangorestframework>=3.16",
    "drf-spectacular>=0.29",
    "gunicorn>=23.0",
    "hiredis>=3.3",
    "httpx[http2]>=0.28",
    "pgvector>=0.4",
    "psycopg[binary]>=3.3",
    "python-dotenv>=1.2",
    "redis>=7.1",
    "resend>=2.27",
    "sentry-sdk[django]>=2.57",
    "stripe>=14.1",
    "structlog>=25.5",
    "taskiq-redis>=1.2",
    "twilio>=9.9",
    "uvicorn[standard]>=0.40",
]

[project.optional-dependencies]
dev = [
    "django-stubs>=5.2",
    "django-stubs-ext>=5.2",
    "markdown>=3.10",
    "pygments>=2.20",
    "pytest>=9.0",
    "pytest-django>=4.11",
    "pytest-mock>=3.15",
    "taskiq[reload]>=0.12",
    "whitenoise>=6.12",
]

[tool.uv]
exclude-newer = "1 week"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "config",
    "contact",
    "payment",
    "sites",
    "teams",
    "users",
]

[tool.hatch.build.targets.wheel.force-include]
"static" = "static"
"templates" = "templates"

[tool.ruff]
target-version = "py314"
exclude = [
    "*/migrations/*",
]

[tool.ruff.lint]
future-annotations = false
select = ["ALL"]
fixable = ["ALL"]
extend-safe-fixes = [
    "EM",    # When raising an f-string exception, move to msg first.
]
ignore = [
    "D",       # Don't enforce docstrings
    "ANN",     # Don't require type annotations
    "TD",      # No rules for TODO comments
    "C901",    # OPTIONAL: We should let Cursor write complex code
    "ARG001",  # Unused arguments are common in DRF and renaming could lead to inheritance problems.
    "ARG002",  # Same as ARG001 but for methods.
    "PLR0913", # We should let Cursor write functions with many arguments.
    "E501",    # Netpick about long lines.
    "F821",    # Undefined names will be caught by Pylance with a suggestion to import the name (more useful than Ruff which doesn't suggest an import action).
    "TRY003",  # We should let Cursor raise descriptive errors without making new Exception classes all the time. Re-using exceptions/messages is not that important.
    "ERA001",  # Commented out code can be useful to leave behind alternative ideas
    # "PLR0912", # We should let Cursor write complex code
    "G004",    # Logging f-strings isn't actually a problem anymore: https://github.com/pylint-dev/pylint/issues/2395
    "TRY301",  # Prohibits a common pattern (raise within try)
    "TRY300",  # Prohibits a common pattern (returning within try)
    "RET504",  # Assigning to a variable before returning is useful for logging of results
    "PLR2004", # Let Cursor make comparisons with hard-coded values.
    "RUF012",  # Enforcing no class attributes doesn't play well with Django Meta.
    "PGH003",  # Takes too long to track down the specific rule codes when ignoring linter rules inline.
    "N806",    # It's nice to have uppercase variable names when they are types.
    "PLR0915", # It's fine to have long function bodies,
    "W293",    # Let blank lines have whitespace in AI system prompt blocks,
    "COM812",  # No need for trailing comma in argument lists,
    "PGH004",  # I want to write 'noqa' in comments sometimes without a specific rule code.
    "B027",    # This rule makes no sense if an abstract class has some no-op-by-default methods.
    "RET505",  # I simply don't agree that this makes the code more readable.
    "PLR6301", # We allow methods that don't use self. These could be useful for subclasses to override. More refactoring work is needed if not.
    "CPY001",  # Don't need copyright at the top of every file.
    "S608",    # Has to do with SQL, which we never use.
    "I002",    # from __future__ import annotations is being deprecated.
    "TC002",   # TYPE_CHECKING is not used in 3.14
]

[tool.ruff.lint.per-file-ignores]
"models.py" = ["PLC0415", "FBT002"]  # Import tasks locally, allow signature of save()
"views.py" = ["PLC0415"]  # Import tasks locally
"*/tests.py" = ["S101", "T201", "INP001"]  # Pytest asserts and prints allowed in test modules
"*/tests/*" = ["S101", "T201", "INP001"]  # Asserts, prints allowed

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--ignore=.git --ignore=data"
markers = [
    "exploratory: mark a test as exploratory (run manually, not part of automated test suite)",
]
